requirements.txt
eel==1.0.0a1
Pillow==8.0.1
xeger==0.3.5